// frontend/src/features/system/types/system.types.ts
export interface SystemNodeScript {
  id: string;
  name: string;
  description?: string;
  variety: 'build' | 'init' | 'sync' | 'custom';
  data?: string;
  enabled: boolean;
  public: boolean;
  created_at: string;
  updated_at: string;
}